+++ demo/app/components/arc-layer/arc-layer.component.ts
@@ -1,8 +1,8 @@
 import { Component, AfterViewInit, OnInit, ViewChild } from '@angular/core';
 import { Observable } from 'rxjs';
-import { ActionType } from '../../../../src/models/action-type.enum';
-import { AcNotification } from '../../../../src/models/ac-notification';
-import { AcLayerComponent } from '../../../../src/components/ac-layer/ac-layer.component';
+import { ActionType } from '../../../../src/angular-cesium/models/action-type.enum';
+import { AcNotification } from '../../../../src/angular-cesium/models/ac-notification';
+import { AcLayerComponent } from '../../../../src/angular-cesium/components/ac-layer/ac-layer.component';
 
 @Component({
 	selector: 'arc-layer',