+++ demo/app/components/dynamic-polyline-layer/dynamic-polyline-layer.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { AcNotification } from '../../../../src/models/ac-notification';
-import { AcLayerComponent } from '../../../../src/components/ac-layer/ac-layer.component';
+import { AcNotification } from '../../../../src/angular-cesium/models/ac-notification';
+import { AcLayerComponent } from '../../../../src/angular-cesium/components/ac-layer/ac-layer.component';
 import { TracksDataProvider } from '../../../utils/services/dataProvider/tracksDataProvider.service';
 
 @Component({