+++ src/angular-cesium-entities-editor/models/edit-mode.enum.ts
@@ -1,0 +1,5 @@
+export enum EditModes {
+  CREATE,
+  EDIT,
+  CREATE_OR_EDIT
+}